Richa_Mishra/complete_project/requirements.txt
faiss-cpu
numpy
sentence-transformers
transformers
torch
pdfplumber
pytesseract
pdf2image
pillow
google-generativeai
python-dotenv
nltk
tqdm